consolidate get_short_hash functions
