+++ src/eduid/common/utils.py
@@ -1,5 +1,7 @@
 __author__ = "lundberg"
 
+from uuid import uuid4
+
 from bson import ObjectId
 
 
@@ -46,3 +48,7 @@ def removesuffix(s: str, suffix: str) -> str:
 
 def make_etag(version: ObjectId):
     return f'W/"{version}"'
+
+
+def get_short_hash(entropy: int = 10) -> str:
+    return uuid4().hex[:entropy]
