+++ src/eduid/maccapi/util.py
@@ -11,10 +11,6 @@ from eduid.maccapi.config import MAccApiConfig
 logger = logging.getLogger(__name__)
 
 
-def get_short_hash(entropy=8):
-    return uuid4().hex[:entropy]
-
-
 def generate_password(length: int = 12) -> str:
     password = pwgen(int(length), no_capitalize=True, no_symbols=True)
     password = " ".join([password[i * 4 : i * 4 + 4] for i in range(0, math.ceil(len(password) / 4))])
