+++ src/eduid/webapp/reset_password/helpers.py
@@ -9,7 +9,7 @@ from flask_babel import gettext as _
 from eduid.common.config.base import EduidEnvironment
 from eduid.common.misc.timeutil import utc_now
 from eduid.common.rpc.exceptions import MailTaskFailed
-from eduid.common.utils import urlappend
+from eduid.common.utils import get_short_hash, urlappend
 from eduid.userdb.exceptions import UserDoesNotExist
 from eduid.userdb.logs import MailAddressProofing, PhoneNumberProofing
 from eduid.userdb.reset_password import ResetPasswordEmailAndPhoneState, ResetPasswordEmailState, ResetPasswordUser
@@ -18,13 +18,7 @@ from eduid.userdb.user import User
 from eduid.webapp.common.api.exceptions import ThrottledException
 from eduid.webapp.common.api.helpers import send_mail
 from eduid.webapp.common.api.messages import FluxData, TranslatableMsg, error_response, success_response
-from eduid.webapp.common.api.utils import (
-    check_password_hash,
-    get_short_hash,
-    get_unique_hash,
-    get_zxcvbn_terms,
-    save_and_sync_user,
-)
+from eduid.webapp.common.api.utils import check_password_hash, get_unique_hash, get_zxcvbn_terms, save_and_sync_user
 from eduid.webapp.common.api.validation import is_valid_password
 from eduid.webapp.common.authn import fido_tokens
 from eduid.webapp.common.authn.utils import generate_password
